+++ crates/vertigo/src/driver_module/js_value/mod.rs
@@ -1,6 +1,5 @@
+mod js_json_list_decoder;
 mod js_json_struct;
-mod js_value_list_decoder;
-mod js_value_struct;
 mod memory_block;
 mod memory_block_read;
 mod memory_block_write;
@@ -8,9 +7,11 @@ mod serialize;
 #[cfg(feature = "chrono")]
 mod serialize_chrono;
 
+pub use js_json_list_decoder::JsJsonListDecoder;
 pub use js_json_struct::{JsJson, JsJsonNumber};
-pub use js_value_struct::JsValue;
 pub use memory_block::MemoryBlock;
+pub use memory_block_read::MemoryBlockRead;
+pub use memory_block_write::MemoryBlockWrite;
 
 #[cfg(test)]
 mod tests;