+++ crates/vertigo/src/driver_module/js_value/tests.rs
@@ -3,15 +3,23 @@ use std::collections::BTreeMap;
 use crate::driver_module::js_value::js_json_struct::JsJsonNumber;
 
 use super::js_json_struct::JsJson;
-use super::js_value_struct::JsValue;
+use crate::{MemoryBlock, MemoryBlockWrite};
+
+fn to_block(json: JsJson) -> MemoryBlock {
+    let size = json.get_size();
+    let block = MemoryBlock::new(size);
+    let mut block_write = MemoryBlockWrite::new(block);
+    json.write_to(&mut block_write);
+    block_write.get_block()
+}
 
 #[test]
 fn json_json_string() {
-    let data1 = JsValue::Json(JsJson::String("test string".into()));
+    let data1 = JsJson::String("test string".into());
 
-    let block = data1.to_block();
+    let block = to_block(data1.clone());
 
-    let Ok(data2) = JsValue::from_block(block) else {
+    let Ok(data2) = JsJson::from_block(block) else {
         unreachable!();
     };
 
@@ -20,22 +28,20 @@ fn json_json_string() {
 
 #[test]
 fn json_json_float() {
-    let data1 = JsValue::F64(3.15);
+    let data1 = JsJson::Number(JsJsonNumber(3.15));
 
-    let block = data1.to_block();
+    let block = to_block(data1.clone());
 
-    let Ok(data2) = JsValue::from_block(block) else {
+    let Ok(data2) = JsJson::from_block(block) else {
         unreachable!();
     };
 
     assert_eq!(data2, data1);
-
-    assert_eq!(data2, JsValue::F64(3.15));
 }
 
 #[test]
 fn json_json_list() {
-    let data1 = JsValue::Json(JsJson::List(vec![
+    let data1 = JsJson::List(vec![
         JsJson::String("aaaa".into()),
         JsJson::String("bbbb".into()),
         JsJson::True,
@@ -50,11 +56,11 @@ fn json_json_list() {
             ("aaa".to_string(), JsJson::Number(JsJsonNumber(2.0))),
             ("bbb".to_string(), JsJson::String(String::from("ccc"))),
         ])),
-    ]));
+    ]);
 
-    let block = data1.to_block();
+    let block = to_block(data1.clone());
 
-    let Ok(data2) = JsValue::from_block(block) else {
+    let Ok(data2) = JsJson::from_block(block) else {
         unreachable!();
     };